Move endpoint tests under routes
